# web/openapi/raffa-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Raffa API",
    "version": "1.1.0",
...
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative to the runtime-configured API origin (AppConfig.apiBaseUrl, ADR-012 \"config, not code\"; see web/src/config/appConfig.ts). This document intentionally does not hard-code a dev/demo host."
    }
  ],
  "paths": {
...
    "/api/market/records/{id}": {
      "get": {
        "operationId": "getMarketRecord",
        "summary": "One market-intelligence record, for the Ask citation side panel.",
        "description": "R-EVD-02 / requirements.md §6 API table: 'one market record (for the citation panel)'; parent story us-01-market-intelligence AC-5: 'returns one record with its provenance label and updatedAt for the citation panel'. No X-Tenant-Id: the market index is shared and read-only for every tenant (ADR-024 'three sources, one rule'), the same 'no authorization boundary to enforce' reasoning GET /api/capabilities above shares. Task E13/F06/US01/T01 (ask-engine) is this route's first writer (backend/src/Raffa.Api/MarketEndpointExtensions.cs); as of this task's own commit that file does not compile (two competing method bodies concatenated mid-statement -- a dangling `if` referencing an out-of-scope `endpoints` variable, a duplicate `deal` local, ~11 duplicate anonymous-object property names), which breaks the whole Raffa.Api assembly, and no Raffa.Api.Tests file exists for this route either. Rather than block this web task on a backend defect outside its own 'Files to create or modify' (backend/**), this operation is authored directly from the task's own coding-objective field list ('title, category, geography, band, provenance label, updatedAt') plus the real, clean, unbroken domain types it is built from: Raffa.Market.Contracts.MarketDeal (record shape), Raffa.Market.Retrieval.MarketNoteComposer.Compose (title/category/geography/provenance/updatedAt -- BuildTitle: '{Supplier} · {Product}'), Raffa.Market.Contracts.MarketProvenance.Label ('representative market data · mock feed · updated yyyy-MM-dd'). `band` is this task's own addition, not named on MarketNote -- the three real percentile prices (MarketDeal.UnitPriceP25/P50/P75) plus Currency, structured rather than left as the citation card's own prose snippet (which already exists separately on the citation, ReplyCitation.Snippet). Once F02/T02's persisted `market_record` store and a fixed MarketEndpointExtensions.cs land, whichever lands last should diff its real response against this document rather than assume a silent match (the same rule this document's own info.description already states for Documents V2).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "MarketDeal.RecordId."
          }
        ],
        "responses": {
          "200": {
            "description": "The record.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["recordId", "title", "category", "geography", "band", "provenance", "updatedAt"],
                  "properties": {
                    "recordId": { "type": "string" },
                    "title": { "type": "string", "description": "MarketNoteComposer.BuildTitle: '{Supplier} · {Product}', or '{Supplier} · {Product} ({Sku})' when a SKU is recorded." },
                    "category": { "type": "string" },
                    "geography": { "type": "string" },
                    "band": {
                      "type": "object",
                      "required": ["p25", "p50", "p75", "currency"],
                      "properties": {
                        "p25": { "type": "number" },
                        "p50": { "type": "number" },
                        "p75": { "type": "number" },
                        "currency": { "type": "string" }
                      }
                    },
                    "provenance": { "type": "string", "description": "MarketProvenance.Label(deal), e.g. 'representative market data · mock feed · updated 2026-09-01'." },
                    "updatedAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No market record with this id."
          }
        }
      }
    },
...
  }
}